+++ .gitignore
@@ -12,3 +12,4 @@ venv/
 
 # IDE
 .vscode/
+.claude/
