+++ backend/config.py
@@ -1,0 +1,33 @@
+import os
+
+class Config:
+    # General configuration
+    SECRET_KEY                  = os.getenv('SECRET_KEY', "una-clave-secreta-de-respaldo")
+    BACKEND_PORT                = int(os.getenv('BACKEND_PORT', 2222))
+    GROQ_API_KEY                = os.getenv('GROQ_API_KEY')
+
+    # Database configuration
+    DB_USER                     = os.getenv('DB_USER', 'OpenWAdmin')
+    DB_PASSWORD                 = os.getenv('DB_PASSWORD', 'Apex_999')
+    DB_NAME                     = os.getenv('DB_NAME', 'OpenW')
+    DB_HOST                     = os.getenv('DB_HOST', 'db')
+    DB_PORT                     = int(os.getenv('DB_PORT', 5432))
+    SQLALCHEMY_DATABASE_URI     = f"postgresql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
+
+    # Facebook Messenger API configuration
+    ACCESS_TOKEN                = os.getenv('ACCESS_TOKEN')
+    APP_ID                      = os.getenv('APP_ID')
+    VERIFY_TOKEN                = os.getenv('VERIFY_TOKEN')
+
+    # Whatsapp API configuration ????
+    WHATSAPP_APP_ID             = os.getenv('WHATSAPP_APP_ID')
+    WHATSAPP_TOKEN              = os.getenv('WHATSAPP_TOKEN')
+    WHATSAPP_PHONE_NUMBER_ID    = os.getenv('WHATSAPP_PHONE_NUMBER_ID')
+    WHATSAPP_BUSINESS_ACCOUNT_ID= os.getenv('WHATSAPP_BUSINESS_ACCOUNT_ID')
+    WHATSAPP_VERIFY_TOKEN       = os.getenv('WHATSAPP_VERIFY_TOKEN')
+
+class DevelopmentConfig(Config):
+    DEBUG = True
+
+class ProductionConfig(Config):
+    DEBUG = False
