+++ backend/db/models/__init__.py
@@ -1,0 +1,15 @@
+from .enums import SeveridadEnum, EstadoEnum
+from .usuarios import Usuarios
+from .vicepresidencia import Vicepresidencia
+from .direccion import Direccion
+from .categoria_novedad import CategoriaNovedad
+from .novedad import Novedad
+from .ia_model import IAModel
+from .configuracion import configuracion
+
+__all__ = [
+    'SeveridadEnum', 'EstadoEnum',
+    'Usuarios', 'Vicepresidencia', 'Direccion',
+    'CategoriaNovedad', 'Novedad',
+    'IAModel', 'configuracion',
+]
