+++ backend/db/models/direccion.py
@@ -1,0 +1,11 @@
+from backend.extensions import db
+from datetime import datetime, timezone
+
+
+class Direccion(db.Model):
+    __tablename__ = 'direccion'
+    id = db.Column(db.Integer, primary_key=True)
+    nombre = db.Column(db.String(150), nullable=False, unique=True)
+    descripcion = db.Column(db.Text, nullable=True)
+    fk_id_vicepresidencia = db.Column(db.Integer, db.ForeignKey('vicepresidencia.id'), nullable=True)
+    created_at = db.Column(db.DateTime, default=lambda: datetime.now(timezone.utc))
