+++ backend/db/schemas/user_dto.py
@@ -1,0 +1,17 @@
+from marshmallow import Schema, fields
+
+
+class UserSchema(Schema):
+    id = fields.Int(dump_only=True)
+    phone = fields.Int()
+    name = fields.Str(required=True)
+    email = fields.Email(required=True)
+    created_at = fields.DateTime(dump_only=True)
+    updated_at = fields.DateTime(dump_only=True)
+    is_admin = fields.Boolean()
+    fk_id_vicepresidencia = fields.Int()
+    fk_id_direccion = fields.Int()
+
+
+user_schema = UserSchema()
+users_schema = UserSchema(many=True)
