+++ backend/db/schemas/vicepresidencia_dto.py
@@ -1,0 +1,16 @@
+from marshmallow import Schema, fields, validate
+
+
+class VicepresidenciaSchema(Schema):
+    id = fields.Int(dump_only=True)
+    nombre = fields.Str(
+        required=True,
+        validate=validate.Length(min=3, max=150),
+        error_messages={"required": "El nombre de la vicepresidencia es obligatorio."},
+    )
+    descripcion = fields.Str(validate=validate.Length(max=500), allow_none=True)
+    created_at = fields.DateTime(dump_only=True)
+
+
+vicepresidencia_schema = VicepresidenciaSchema()
+vicepresidencias_schema = VicepresidenciaSchema(many=True)
