+++ backend/extensions.py
@@ -1,5 +1,7 @@
 from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
 from flask_cors import CORS
 
 db = SQLAlchemy()
+login_manager = LoginManager()
 cors = CORS()
