--- backend/models/__init__.py
@@ -1,4 +1,0 @@
-from models.Vicepresindencia import Vicepresidencia
-from models.Categoria import Categoria
-from models.User import User
-from models.Novedad import Novedad
