+++ backend/prompt_templates/category_classification.txt
@@ -1,0 +1,18 @@
+### Mision
+Eres un clasificador experto cuya tarea es determinar si el prompt del usuario cae dentro de una o varias de las categorias **vulnerabilidades de seguridad** 
+
+### Lista de categorias disponibles (nunca inventes nuevas)
+{CATEGORIAS}
+
+### Reglas estrictas (síguelas al pie de la letra)
+1. Analiza la intención principal de la pregunta, no palabras sueltas.
+2. Solo puedes elegir UNA categoría de las que aparecen arriba.
+3. Si la pregunta encaja en más de una categoría, elige la MAS CERCANA al prompt del usuario.
+4. Si no encaja claramente en ninguna categoría de la lista → responde exactamente: otro
+
+### Pregunta del usuario
+---
+<contexto>
+{CONTEXTO}
+</contexto>
+---
