+++ backend/requirements.txt
@@ -1,5 +1,6 @@
 flask
 flask-cors
+flask-login
 flask-sqlalchemy
 psycopg2-binary
 pgvector
