+++ docker-compose.yml
@@ -11,6 +11,7 @@ services:
     environment:
       - FLASK_APP=app:create_app
       - FLASK_ENV=${FLASK_MODE}
+    command: /bin/sh -c "dos2unix /app/entrypoint.sh && sh /app/entrypoint.sh"
     depends_on:
       db:
         condition: service_healthy
@@ -42,7 +43,7 @@ services:
     ports:
       - "${DB_EXPOSED_PORT}:5432"
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U ${DB_USER} -d ${DB_NAME}"]
+      test: [ "CMD-SHELL", "pg_isready -U ${DB_USER} -d ${DB_NAME}" ]
       interval: 5s
       timeout: 5s
       retries: 5
